# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Campaign Nexus",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/campaigns/{campaign_id}/npcs": {
      "get": {
        "tags": [
          "npcs"
        ],
        "summary": "List Npcs",
        "description": "The saved queries (FR-6.6): by status, location, faction, met-party, knows-X.",
        "operationId": "list_npcs_api_v1_campaigns__campaign_id__npcs_get",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "location_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Location Id"
            }
          },
          {
            "name": "faction_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Faction Id"
            }
          },
          {
            "name": "met_party",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Met Party"
            }
          },
          {
            "name": "knows",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Knows"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NpcOut"
                  },
                  "title": "Response List Npcs Api V1 Campaigns  Campaign Id  Npcs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "npcs"
        ],
        "summary": "Create Npc",
        "operationId": "create_npc_api_v1_campaigns__campaign_id__npcs_post",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NpcCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NpcOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaign_id}/npcs/schedules": {
      "get": {
        "tags": [
          "npcs"
        ],
        "summary": "List All Schedules",
        "operationId": "list_all_schedules_api_v1_campaigns__campaign_id__npcs_schedules_get",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScheduleOut"
                  },
                  "title": "Response List All Schedules Api V1 Campaigns  Campaign Id  Npcs Schedules Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaign_id}/npcs/schedules/{schedule_id}": {
      "delete": {
        "tags": [
          "npcs"
        ],
        "summary": "Delete Schedule",
        "operationId": "delete_schedule_api_v1_campaigns__campaign_id__npcs_schedules__schedule_id__delete",
        "parameters": [
          {
            "name": "schedule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Schedule Id"
            }
          },
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaign_id}/npcs/{npc_id}": {
      "get": {
        "tags": [
          "npcs"
        ],
        "summary": "Get Npc",
        "operationId": "get_npc_api_v1_campaigns__campaign_id__npcs__npc_id__get",
        "parameters": [
          {
            "name": "npc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Npc Id"
            }
          },
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NpcOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "npcs"
        ],
        "summary": "Update Npc",
        "operationId": "update_npc_api_v1_campaigns__campaign_id__npcs__npc_id__patch",
        "parameters": [
          {
            "name": "npc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Npc Id"
            }
          },
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NpcUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NpcOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaign_id}/npcs/{npc_id}/relocate": {
      "post": {
        "tags": [
          "npcs"
        ],
        "summary": "Relocate",
        "operationId": "relocate_api_v1_campaigns__campaign_id__npcs__npc_id__relocate_post",
        "parameters": [
          {
            "name": "npc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Npc Id"
            }
          },
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RelocateIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NpcOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaign_id}/npcs/{npc_id}/status": {
      "post": {
        "tags": [
          "npcs"
        ],
        "summary": "Set Status",
        "operationId": "set_status_api_v1_campaigns__campaign_id__npcs__npc_id__status_post",
        "parameters": [
          {
            "name": "npc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Npc Id"
            }
          },
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NpcOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaign_id}/npcs/{npc_id}/interactions": {
      "post": {
        "tags": [
          "npcs"
        ],
        "summary": "Record Interaction",
        "operationId": "record_interaction_api_v1_campaigns__campaign_id__npcs__npc_id__interactions_post",
        "parameters": [
          {
            "name": "npc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Npc Id"
            }
          },
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InteractionIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NpcOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaign_id}/npcs/{npc_id}/history": {
      "get": {
        "tags": [
          "npcs"
        ],
        "summary": "Get History",
        "operationId": "get_history_api_v1_campaigns__campaign_id__npcs__npc_id__history_get",
        "parameters": [
          {
            "name": "npc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Npc Id"
            }
          },
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HistoryRow"
                  },
                  "title": "Response Get History Api V1 Campaigns  Campaign Id  Npcs  Npc Id  History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaign_id}/npcs/{npc_id}/where": {
      "get": {
        "tags": [
          "npcs"
        ],
        "summary": "Where Was",
        "description": "\"Where is X now / where was X at time T / where was X during session N\" (FR-6.2).\n\nA session is a *span* of campaign time, so it answers with every place the NPC occupied\nwhile it ran; an instant answers with exactly one.",
        "operationId": "where_was_api_v1_campaigns__campaign_id__npcs__npc_id__where_get",
        "parameters": [
          {
            "name": "npc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Npc Id"
            }
          },
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          },
          {
            "name": "at_game",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "At Game"
            }
          },
          {
            "name": "session_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhereOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaign_id}/npcs/{npc_id}/schedules": {
      "get": {
        "tags": [
          "npcs"
        ],
        "summary": "List Schedules",
        "operationId": "list_schedules_api_v1_campaigns__campaign_id__npcs__npc_id__schedules_get",
        "parameters": [
          {
            "name": "npc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Npc Id"
            }
          },
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScheduleOut"
                  },
                  "title": "Response List Schedules Api V1 Campaigns  Campaign Id  Npcs  Npc Id  Schedules Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "npcs"
        ],
        "summary": "Create Schedule",
        "operationId": "create_schedule_api_v1_campaigns__campaign_id__npcs__npc_id__schedules_post",
        "parameters": [
          {
            "name": "npc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Npc Id"
            }
          },
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NpcCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "alive"
          },
          "location_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Id"
          },
          "goals": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Goals"
          },
          "secrets": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Secrets"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "NpcCreate"
      },
      "NpcOut": {
        "properties": {
          "entity_id": {
            "type": "string",
            "title": "Entity Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "current_location_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Location Id"
          },
          "current_location_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Location Name"
          },
          "has_met_party": {
            "type": "boolean",
            "title": "Has Met Party"
          },
          "last_party_interaction_game": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Party Interaction Game"
          },
          "goals": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Goals"
          },
          "secrets": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Secrets"
          },
          "voice_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voice Notes"
          },
          "knows_about": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Knows About"
          }
        },
        "type": "object",
        "required": [
          "entity_id",
          "name",
          "summary",
          "status",
          "current_location_id",
          "current_location_name",
          "has_met_party",
          "last_party_interaction_game",
          "goals",
          "secrets",
          "voice_notes",
          "knows_about"
        ],
        "title": "NpcOut"
      },
      "NpcUpdate": {
        "properties": {
          "goals": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Goals"
          },
          "secrets": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Secrets"
          },
          "voice_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voice Notes"
          }
        },
        "type": "object",
        "title": "NpcUpdate"
      },
...
      "ScheduleOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "npc_id": {
            "type": "string",
            "title": "Npc Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "interval_days": {
            "type": "integer",
            "title": "Interval Days"
          },
          "stops": {
            "items": {
              "$ref": "#/components/schemas/ScheduleStop"
            },
            "type": "array",
            "title": "Stops"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "materialized_through_game": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Materialized Through Game"
          }
        },
        "type": "object",
        "required": [
          "id",
          "npc_id",
          "label",
          "interval_days",
          "stops",
          "active",
          "materialized_through_game"
        ],
        "title": "ScheduleOut"
      },
...
    }
  }
}